Work/monster-skill-analyzer/requirements.txt
flask>=3.0.0
flask-wtf>=1.2.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.0.0
pyyaml>=6.0.1
pandas>=2.0.0
